Share number might be larger than long
Total share number might overflow long max value, even if the database
number which is typically divided by 1000 doesn't. This patch changes
the database column "value" to be BIGINT, which might be overkill and we
may decide to go back to INT. At the same time does the internal
computations with long long and import with setInt64.

Also return 0 for DY when 7.08.04.01 does not have any value as is the
case in financial companies for example.

Fixes #5

+++ src/dfp/dfp_cvm_file.cpp
@@ -343,7 +343,7 @@ void Dfp::CvmFile::import (const Dfp::Database &conn) {
         xmldoc.child("ArrayOfComposicaoCapitalSocialDemonstracaoFinanceira").\
         child("ComposicaoCapitalSocialDemonstracaoFinanceira").\
         child_value(share_number_string[i].c_str() ) );
-    acct.value  = (int) acct.value / quant_scale;
+    acct.value  = acct.value / quant_scale;
     acct.balance_type = DFP_BALANCE_IF;
     acct.financial_info_type = DFP_FINANCIAL_INFO_CONSOLIDATED;
     acct.number = share_number_acct[i];
